[openrndr-gl3] Add configuration options that control the extraction of angle libraries

+++ openrndr-jvm/openrndr-gl3/src/jvmMain/kotlin/org/openrndr/internal/gl3/angle/Angle.kt
@@ -1,7 +1,7 @@
 package org.openrndr.internal.gl3.angle
 
 import io.github.oshai.kotlinlogging.KotlinLogging
-import org.openrndr.platform.Platform
+import org.openrndr.internal.gl3.DriverGL3Configuration
 import java.io.File
 
 private class Angle
@@ -10,16 +10,30 @@ private val logger = KotlinLogging.logger { }
 fun extractAngleLibraries() {
     logger.info { "Loading ANGLE libraries from resources" }
     val targetDirectory = File(".")// Platform.tempDirectory()
-    val egl = Angle::class.java.getResourceAsStream("/org/openrndr/internal/gl3/angle/libEGL.dylib") ?: error("libEGL.dylib not found in resources")
-    File(targetDirectory,"libEGL.dylib").outputStream().use {
-        egl.copyTo(it)
+    val egl = Angle::class.java.getResourceAsStream("/org/openrndr/internal/gl3/angle/libEGL.dylib")
+        ?: error("libEGL.dylib not found in resources")
+    val targetEgl = File(targetDirectory, "libEGL.dylib")
+
+    if (!targetEgl.exists() || DriverGL3Configuration.overwriteExistingAngle) {
+        targetEgl.outputStream().use {
+            egl.copyTo(it)
+        }
     }
 
-    val glesv2 = Angle::class.java.getResourceAsStream("/org/openrndr/internal/gl3/angle/libGLESv2.dylib") ?: error("libGLESv2.dylib not found in resources")
-    File(targetDirectory,"libGLESv2.dylib").outputStream().use {
-        glesv2.copyTo(it)
+
+    val targetGlesv2 = File(targetDirectory, "libGLESv2.dylib")
+    if (!targetGlesv2.exists() || DriverGL3Configuration.overwriteExistingAngle) {
+        val glesv2 = Angle::class.java.getResourceAsStream("/org/openrndr/internal/gl3/angle/libGLESv2.dylib") ?: error(
+            "libGLESv2.dylib not found in resources"
+        )
+        targetGlesv2.outputStream().use {
+            glesv2.copyTo(it)
+        }
     }
-    File(targetDirectory, "libGLESv2.dylib").deleteOnExit()
-    File(targetDirectory, "libEGL.dylib").deleteOnExit()
+
+    if (DriverGL3Configuration.deleteAngleOnExit) {
+        File(targetDirectory, "libGLESv2.dylib").deleteOnExit()
+        File(targetDirectory, "libEGL.dylib").deleteOnExit()
+    }
 }
